reworked cinder stats calculations
After a lot of research into cinder's pool stats and how each are used
by the cinder capacity filter and capacity weighter, I discovered how
inconsistent cinder is in it's algorithm for determining available
space for provisioning requests.

This patch reflects a more consistent and accurate algorithm for how
each field is used as defined by the cinder documentation and intentions
for each field.  Such as reserved percent, allocated,
provisioned_capacity, etc.

https://specs.openstack.org/openstack/cinder-specs/specs/queens/provisioning-improvements.html

This patch now reports virtual_free_capacity_gb, which is a combination
of the total capacity reported, the reserved percentage applied to the
total capacity (called available_capacity) and the
max_over_subscription_ratio if thin provisioning is enabled.
As the allocated capacity approaches the reserved percentage and over
subscription ratio, that virtual_free_capacity_gb will converge to 0.
So there is no need to look at available intil overcommit is reached or
available until reserved is met.   The virtual_free_capacity accounts
for all of it.