Use unified calculate_capacity_factors.
This patch adds the new cinder function calculate_capacity_factors.
This method is being added upstream to unify the way cinder calculates
the virtual free space available against a pool.  This also builds
a dictionary of all of the factors that are used to calculate that
virtual free space.   Those factors are also what hammer os poool-stats
reports for each pool(datastore).   Eventually the factors dictionary
will be added to the pools reporting by the scheduler, so cinder will
provide the definitive answer for the factors.  For now, we embed
the new function in hammer.