[Cinder] refactor the Cinder collector a bit and added backend_state
This patch cleans up the logging of the stats collected.
This combines the creation of the Gauge that is yielded to the caller
and the logging of the Gauge info to the log in 1 place for each
stat collected.  This will make it easier to add new items in the
future.

This patch also adds a new stat of backend_state which can be 'up' or
'down'.   Once the new cinder patch is available it will be easier
to add the new pool_state for when datastores go into maintenance mode.

This patch also adds the new pool_state, which denotes if a particular
datastore/pool is available for cinder to provision against.