Support loading secrets from different config file
Deploying the openstack-exporter via k8s, we want to split up config and
secrets into ConfigMap and Secret k8s objects. To achieve this, we have
to support loading username and password from another file. Therefore,
we add a new option `--secrets` that requires a YAML file path. An
example of such a YAML file is included in this commit.