[Cinder] Allow only expected backends
This patch updates the cinderbackend exporter to honor the
custom config settings for expected_backends combined with
the allow_unexpected_backends.

This will allow us to only export stats for enabled backends
in regions.  Such as we have regions that have vmware_fcd enabled
and other regions that don't have it enabled.  When we don't have
a backend enabled, the exporter reports that the backend is down,
which turns into a critical alert.

We are also in the process of disabling vmware backends in the regions
where we have enabled vmware_fcd and migrated everything off of vmware
to vmware_fcd.